divide login, people. project
