+++ src/components/header/header.jsx
@@ -1,33 +1,18 @@
-import React from 'react';
-import styles from './header.module.css';
-import { GoogleLogin } from 'react-google-login';
+import React from "react";
+import styles from "./header.module.css";
+import Login from "../login/login";
+import People from "../people/people";
+import Project from "../project/project";
 const Header = (props) => {
-	const responseGoogle = (res) => {
-		console.log(res);
-	};
-
-	const responseFail = (err) => {
-		console.log(err);
-	};
-
-	return (
-		<header className={styles.header}>
-			<img className={styles.logoImage} src="/images/logo.png" alt="logo" />
-			<h1 className={styles.logo}>RCV</h1>
-			<div className={styles.container}>
-				<a className={styles.login}>Login</a>
-				<a className={styles.login}>Add a new project</a>
-				<GoogleLogin
-					className={styles.googleLogin}
-					clientId={process.env.REACT_APP_GOOGLE_API_KEY}
-					buttonText="Sign in with google"
-					onSuccess={responseGoogle}
-					onFailure={responseFail}
-					cookiePolicy={'single_host_origin'}
-				/>
-			</div>
-		</header>
-	);
+  return (
+    <header className={styles.header}>
+      <img className={styles.logoImage} src="/images/logo.png" alt="logo" />
+      <h1 className={styles.logo}>RCV</h1>
+      <div className={styles.container}>
+        <Login />
+      </div>
+    </header>
+  );
 };
 
 export default Header;
