+++ src/components/people/people.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import styles from "./people.module.css";
+const People = (props) => {
+  return (
+    <div>
+      <div>People</div>
+    </div>
+  );
+};
+
+export default People;
