+++ src/components/project/project.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import styles from "./project.module.css";
+const Project = (props) => {
+  return (
+    <div>
+      <div>Project</div>
+    </div>
+  );
+};
+
+export default Project;
